Refactor code, avoid duplicating code for each online judge
