+++ app/src/Config/Types/LangAliases.ts
@@ -1,9 +1,2 @@
-export type LangAliases = {
-  codeforces?: string;
-  universaloj?: string;
-  ucup?: string;
-  atcoder?: string;
-  omegaup?: string;
-  szkopul?: string;
-  yandex?: string;
-};
+import { OnlineJudgeName } from "./OnlineJudgeName";
+export type LangAliases = { [K in OnlineJudgeName]?: string };
