+++ app/src/Config/Types/OnlineJudgeName.ts
@@ -1,0 +1,9 @@
+export enum OnlineJudgeName {
+  codeforces = "codeforces",
+  universaloj = "universaloj",
+  ucup = "ucup",
+  atcoder = "atcoder",
+  omegaup = "omegaup",
+  szkopul = "szkopul",
+  yandex = "yandex"
+};
