+++ app/src/Submit/OnlineJudgeFactory/Szkopul.ts
@@ -17,7 +17,8 @@
  */
 
 import { Page } from "playwright-chromium";
-import OnlineJudge, { OnlineJudgeName } from "./OnlineJudge";
+import OnlineJudge from "./OnlineJudge";
+import { OnlineJudgeName } from "../../Config/Types/OnlineJudgeName";
 
 export default class Szkopul extends OnlineJudge {
   readonly onlineJudgeName = OnlineJudgeName.szkopul;
